+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Links in docs
 
+### Added
+
+- drawPath can now be used differently and this new way can be described with ts.
+The old style is `drawPath(x1, y1, x2, y2..., options)` we now allow `drawPath([[x1, y1], [x2, y2]...], options)` too
+
 ### Changed
 
 - Bump dev dependency versions s